+++ chain/chain/Cargo.toml
@@ -10,7 +10,6 @@ workspace = true
 
 [dependencies]
 actix.workspace = true
-assert_matches.workspace = true
 borsh.workspace = true
 bytesize.workspace = true
 chrono.workspace = true
