+++ chain/client/src/client.rs
@@ -21,7 +21,6 @@ use near_chain::chain::{
     ApplyChunksDoneMessage, BlockCatchUpRequest, BlockMissingChunks, BlocksCatchUpState,
     VerifyBlockHashAndSignatureResult,
 };
-use near_chain::flat_storage_creator::FlatStorageCreator;
 use near_chain::orphan::OrphanMissingChunks;
 use near_chain::state_snapshot_actor::SnapshotCallbacks;
 use near_chain::test_utils::format_hash;
@@ -185,8 +184,6 @@ pub struct Client {
     tier1_accounts_cache: Option<(EpochId, Arc<AccountKeys>)>,
     /// Resharding sender.
     pub resharding_sender: ReshardingSender,
-    /// Used when it is needed to create flat storage in background for some shards.
-    flat_storage_creator: Option<FlatStorageCreator>,
     /// A map storing the last time a block was requested for state sync.
     pub last_time_sync_block_requested: HashMap<CryptoHash, near_async::time::Utc>,
     /// Helper module for stateless validation functionality like chunk witness production, validation
@@ -278,20 +275,13 @@ impl Client {
             runtime_adapter.clone(),
             &chain_genesis,
             doomslug_threshold_mode,
-            chain_config.clone(),
+            chain_config,
             snapshot_callbacks,
             async_computation_spawner.clone(),
             validator_signer.clone(),
             resharding_sender.clone(),
         )?;
-        // Create flat storage or initiate migration to flat storage.
-        let flat_storage_creator = FlatStorageCreator::new(
-            epoch_manager.clone(),
-            runtime_adapter.clone(),
-            chain.chain_store(),
-            &chain.resharding_manager.flat_storage_resharder.clone(),
-            chain_config.background_migration_threads,
-        )?;
+        chain.init_flat_storage()?;
         let sharded_tx_pool =
             ShardedTransactionPool::new(rng_seed, config.transaction_pool_size_limit);
         let sync_status = SyncStatus::AwaitingPeers;
@@ -405,7 +395,6 @@ impl Client {
             ),
             tier1_accounts_cache: None,
             resharding_sender,
-            flat_storage_creator,
             last_time_sync_block_requested: HashMap::new(),
             chunk_validator,
             chunk_inclusion_tracker: ChunkInclusionTracker::new(),
@@ -2669,19 +2658,6 @@ impl Client {
         //            self.challenges.insert(challenge.hash, challenge);
         //        }
         Ok(())
-    }
-
-    /// Check updates from background flat storage creation processes and possibly update
-    /// creation statuses. Returns boolean indicating if all flat storages are created or
-    /// creation is not needed.
-    pub fn run_flat_storage_creation_step(&mut self) -> Result<bool, Error> {
-        let result = match &mut self.flat_storage_creator {
-            Some(flat_storage_creator) => {
-                flat_storage_creator.update_status(self.chain.chain_store())?
-            }
-            None => true,
-        };
-        Ok(result)
     }
 }
