+++ chain/client/src/client_actor.rs
@@ -911,8 +911,6 @@ impl fmt::Display for SyncRequirement {
 
 impl ClientActorInner {
     pub fn start(&mut self, ctx: &mut dyn DelayedActionRunner<Self>) {
-        self.start_flat_storage_creation(ctx);
-
         // Start syncing job.
         self.start_sync(ctx);
 
@@ -1518,29 +1516,6 @@ impl ClientActorInner {
                 Ok(SyncRequirement::AlreadyCaughtUp { peer_id, highest_height, head })
             }
         }
-    }
-
-    fn start_flat_storage_creation(&mut self, ctx: &mut dyn DelayedActionRunner<Self>) {
-        if !self.client.config.flat_storage_creation_enabled {
-            return;
-        }
-        match self.client.run_flat_storage_creation_step() {
-            Ok(false) => {}
-            Ok(true) => {
-                return;
-            }
-            Err(err) => {
-                error!(target: "client", "Error occurred during flat storage creation step: {:?}", err);
-            }
-        }
-
-        ctx.run_later(
-            "ClientActor start_flat_storage_creation",
-            self.client.config.flat_storage_creation_period,
-            move |act, ctx| {
-                act.start_flat_storage_creation(ctx);
-            },
-        );
     }
 
     /// Starts syncing and then switches to either syncing or regular mode.
