+++ core/o11y/src/metrics.rs
@@ -217,11 +217,6 @@ static EXCEPTIONS: LazyLock<HashSet<&str>> = LazyLock::new(|| {
         "flat_storage_cached_changes_num_items",
         "flat_storage_cached_changes_size",
         "flat_storage_cached_deltas",
-        "flat_storage_creation_fetched_state_items",
-        "flat_storage_creation_fetched_state_parts",
-        "flat_storage_creation_remaining_state_parts",
-        "flat_storage_creation_status",
-        "flat_storage_creation_threads_used",
         "flat_storage_distance_to_head",
         "flat_storage_head_height",
         "flat_storage_hops_to_head",
