+++ core/store/src/flat/mod.rs
@@ -37,7 +37,6 @@ mod types;
 pub use chunk_view::FlatStorageChunkView;
 pub use delta::{FlatStateChanges, FlatStateDelta, FlatStateDeltaMetadata};
 pub use manager::FlatStorageManager;
-pub use metrics::FlatStorageCreationMetrics;
 pub use storage::FlatStorage;
 pub use types::{
     BlockInfo, FetchingStateStatus, FlatStateIterator, FlatStorageCreationStatus, FlatStorageError,
@@ -45,9 +44,3 @@ pub use types::{
 };
 
 pub(crate) const POISONED_LOCK_ERR: &str = "The lock was poisoned.";
-
-/// Number of traversed parts during a single step of fetching state.
-pub const NUM_PARTS_IN_ONE_STEP: u64 = 20;
-
-/// Memory limit for state part being fetched.
-pub const STATE_PART_MEMORY_LIMIT: bytesize::ByteSize = bytesize::ByteSize(10 * bytesize::MIB);
