+++ docs/architecture/storage/flat_storage.md
@@ -196,10 +196,6 @@ It holds all FlatStorages which NightshadeRuntime knows about and:
 Interface for getting ValueRefs from flat storage for some shard for some fixed block, supported
 by get_ref method.
 
-## FlatStorageCreator
-
-Creates flat storages for all tracked shards or initiates process of background flat storage creation if for some shard we have only Trie but not FlatStorage. Supports update_status which checks background job results and updates creation statuses and should create flat storage when all jobs are finished.
-
 ## Other notes
 
 ### Chain dependency
